+++ file_search/__init__.py
@@ -1,0 +1,15 @@
+"""Utilities for indexing and searching files."""
+
+from .database import index_files, choose_directory, update_database
+from .search import search_files
+from .monitor import start_monitoring
+from .gui import FileSearchApp
+
+__all__ = [
+    "index_files",
+    "choose_directory",
+    "update_database",
+    "search_files",
+    "start_monitoring",
+    "FileSearchApp",
+]
